feat: move HoverForm component to a new location and update imports
