+++ src/renderer/components/states/hover/components/HoverContent.tsx
@@ -27,7 +27,7 @@
 import type { ReactElement } from 'react';
 import type { HoverContentProps } from '../config/hoverConfig';
 import { useHover } from '../hooks/useHover';
-import { HoverForm } from '../pages/HoverForm';
+import { HoverForm } from './HoverForm';
 import '../../../styles/hover/hover.css';
 
 /** Hover 状态内容组件 */
